setting up for testing
I have the model file set up for testing now, so it will print some debug of all the files in training_files if run.

midi_randomizer will print out all supporting notes and their ids. Then it will create the midi file for the generated music to be able to be played.